# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartFinance AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AccountType = (typeof AccountType)[keyof typeof AccountType];

export const AccountType = {
  checking: "checking",
  savings: "savings",
  credit: "credit",
  investment: "investment",
  loan: "loan",
} as const;

export interface Account {
  id: number;
  name: string;
  type: AccountType;
  balance: number;
  currency: string;
  institution?: string;
  isActive: boolean;
  createdAt: string;
}

export type CreateAccountBodyType =
  (typeof CreateAccountBodyType)[keyof typeof CreateAccountBodyType];

export const CreateAccountBodyType = {
  checking: "checking",
  savings: "savings",
  credit: "credit",
  investment: "investment",
  loan: "loan",
} as const;

export interface CreateAccountBody {
  name: string;
  type: CreateAccountBodyType;
  balance: number;
  currency?: string;
  institution?: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  income: "income",
  expense: "expense",
  transfer: "transfer",
} as const;

export interface Transaction {
  id: number;
  accountId: number;
  amount: number;
  description: string;
  category: string;
  subcategory?: string;
  type: TransactionType;
  date: string;
  merchant?: string;
  isRecurring: boolean;
  aiCategorized: boolean;
  notes?: string;
  createdAt: string;
}

export type CreateTransactionBodyType =
  (typeof CreateTransactionBodyType)[keyof typeof CreateTransactionBodyType];

export const CreateTransactionBodyType = {
  income: "income",
  expense: "expense",
  transfer: "transfer",
} as const;

export interface CreateTransactionBody {
  accountId: number;
  amount: number;
  description: string;
  category: string;
  subcategory?: string;
  type: CreateTransactionBodyType;
  date: string;
  merchant?: string;
  isRecurring?: boolean;
  notes?: string;
}

export type UpdateTransactionBodyType =
  (typeof UpdateTransactionBodyType)[keyof typeof UpdateTransactionBodyType];

export const UpdateTransactionBodyType = {
  income: "income",
  expense: "expense",
  transfer: "transfer",
} as const;

export interface UpdateTransactionBody {
  amount?: number;
  description?: string;
  category?: string;
  subcategory?: string;
  type?: UpdateTransactionBodyType;
  date?: string;
  merchant?: string;
  isRecurring?: boolean;
  notes?: string;
}

export interface PaginatedTransactions {
  data: Transaction[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type BulkCategorizeResultResultsItem = {
  id: number;
  category: string;
  confidence: number;
};

export interface BulkCategorizeResult {
  categorized: number;
  failed: number;
  results: BulkCategorizeResultResultsItem[];
}

export interface Budget {
  id: number;
  category: string;
  limit: number;
  spent: number;
  month: number;
  year: number;
  alertThreshold?: number;
  createdAt: string;
}

export interface CreateBudgetBody {
  category: string;
  limit: number;
  month: number;
  year: number;
  alertThreshold?: number;
}

export interface UpdateBudgetBody {
  limit?: number;
  alertThreshold?: number;
}

export interface MonthlyTrend {
  month: string;
  income: number;
  expenses: number;
  savings: number;
}

export interface CategoryBreakdown {
  category: string;
  amount: number;
  percentage: number;
  transactionCount: number;
  color: string;
}

export type FinancialSummaryChangeFromLastMonth = {
  income: number;
  expenses: number;
  savings: number;
};

export interface FinancialSummary {
  totalBalance: number;
  monthlyIncome: number;
  monthlyExpenses: number;
  monthlySavings: number;
  savingsRate: number;
  netWorth: number;
  creditScore?: number;
  aiInsight: string;
  changeFromLastMonth: FinancialSummaryChangeFromLastMonth;
}

export interface AIAnalysisRequest {
  question: string;
  context?: string;
}

export interface AIAnalysisResponse {
  analysis: string;
  recommendations: string[];
  score?: number;
}

export type AnomalySeverity =
  (typeof AnomalySeverity)[keyof typeof AnomalySeverity];

export const AnomalySeverity = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Anomaly {
  transactionId: number;
  description: string;
  amount: number;
  category: string;
  date: string;
  severity: AnomalySeverity;
  reason: string;
}

export type TaxSummaryCategoriesItem = {
  category: string;
  amount: number;
  isDeductible: boolean;
};

export interface TaxSummary {
  year: number;
  totalIncome: number;
  taxableIncome: number;
  deductibleExpenses: number;
  estimatedTax: number;
  categories: TaxSummaryCategoriesItem[];
}

export interface Conversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface ConversationWithMessages {
  id: number;
  title: string;
  messages: Message[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export type ListTransactionsParams = {
  page?: number;
  limit?: number;
  category?: string;
  accountId?: number;
  startDate?: string;
  endDate?: string;
  type?: ListTransactionsType;
};

export type ListTransactionsType =
  (typeof ListTransactionsType)[keyof typeof ListTransactionsType];

export const ListTransactionsType = {
  income: "income",
  expense: "expense",
  transfer: "transfer",
} as const;

export type ListBudgetsParams = {
  month?: string;
  year?: number;
};

export type GetSpendingTrendsParams = {
  months?: number;
};

export type GetCategoryBreakdownParams = {
  month?: string;
  year?: number;
};

export type GetTaxSummaryParams = {
  year?: number;
};