# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SmartFinance AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      summary: List all budgets
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create a budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetBody"
      responses:
        "201":
          description: Budget created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
  /budgets/{id}:
    patch:
      operationId: updateBudget
      tags: [budgets]
      summary: Update a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBudgetBody"
      responses:
        "200":
          description: Updated budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /insights/spending-trends:
    get:
      operationId: getSpendingTrends
      tags: [insights]
      summary: Get spending trends by month
      parameters:
        - name: months
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Spending trends
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyTrend"
  /insights/category-breakdown:
    get:
      operationId: getCategoryBreakdown
      tags: [insights]
      summary: Get spending breakdown by category
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryBreakdown"
...
  /insights/ai-analysis:
    post:
      operationId: getAIAnalysis
      tags: [insights]
      summary: Get AI-powered financial analysis and recommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIAnalysisRequest"
      responses:
        "200":
          description: AI analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIAnalysisResponse"
...
  /insights/tax-summary:
    get:
      operationId: getTaxSummary
      tags: [insights]
      summary: Get tax-related transaction summary
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Tax summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxSummary"
...
components:
  schemas:
...
    Budget:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        limit:
          type: number
        spent:
          type: number
        month:
          type: integer
        year:
          type: integer
        alertThreshold:
          type: number
        createdAt:
          type: string
      required: [id, category, limit, spent, month, year, createdAt]
    CreateBudgetBody:
      type: object
      properties:
        category:
          type: string
        limit:
          type: number
        month:
          type: integer
        year:
          type: integer
        alertThreshold:
          type: number
      required: [category, limit, month, year]
    UpdateBudgetBody:
      type: object
      properties:
        limit:
          type: number
        alertThreshold:
          type: number
    MonthlyTrend:
      type: object
      properties:
        month:
          type: string
        income:
          type: number
        expenses:
          type: number
        savings:
          type: number
      required: [month, income, expenses, savings]
    CategoryBreakdown:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        percentage:
          type: number
        transactionCount:
          type: integer
        color:
          type: string
      required: [category, amount, percentage, transactionCount, color]
...
    AIAnalysisRequest:
      type: object
      properties:
        question:
          type: string
        context:
          type: string
      required: [question]
    AIAnalysisResponse:
      type: object
      properties:
        analysis:
          type: string
        recommendations:
          type: array
          items:
            type: string
        score:
          type: number
      required: [analysis, recommendations]
...
    TaxSummary:
      type: object
      properties:
        year:
          type: integer
        totalIncome:
          type: number
        taxableIncome:
          type: number
        deductibleExpenses:
          type: number
        estimatedTax:
          type: number
        categories:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              amount:
                type: number
              isDeductible:
                type: boolean
            required: [category, amount, isDeductible]
      required: [year, totalIncome, taxableIncome, deductibleExpenses, estimatedTax, categories]
...