# lib/api-zod/src/generated/types/taxSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartFinance AI API
 * OpenAPI spec version: 0.1.0
 */
import type { TaxSummaryCategoriesItem } from "./taxSummaryCategoriesItem";

export interface TaxSummary {
  year: number;
  totalIncome: number;
  taxableIncome: number;
  deductibleExpenses: number;
  estimatedTax: number;
  categories: TaxSummaryCategoriesItem[];
}