# lib/api-zod/src/generated/types/updateBudgetBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartFinance AI API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateBudgetBody {
  limit?: number;
  alertThreshold?: number;
}